requirements.txt
Flask>=2.0
mysql-connector-python>=8.0
Werkzeug>=2.0

